requirements.txt
tesseract-ocr
libtesseract-dev
fastapi[all]
opencv-python
tesseract